Clean up + File split
